Revert "Merge remote-tracking branch 'origin/develop' into bitcoin-multisig-trnsBuild"
This reverts commit 6d436266739c0d668864b67ddd5f44b502711bbb, reversing
changes made to 954506dc259d4cc11cc6451154eb2bfc8ea2d0a9.
