--- shared/components/QrReader/styles.scss
@@ -1,19 +1,0 @@
-.scan {
-  display: flex;
-  flex-direction: column;
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: #fff;
-  width: 100%;
-  height: 100%;
-  z-index: 999999999999999999;
-}
-
-.close {
-  text-align: right;
-  position: relative;
-  font-size: 30px;
-}
