+++ shared/components/controls/NewButton/NewButton.js
@@ -32,7 +32,6 @@ const NewButton = (props) => {
       onClick={onClick}
       id={id}
       disabled={disabled}
-      data-tip data-for={id}
     >
       {children}
     </button>
