+++ shared/components/forms/Input/Input.js
@@ -63,7 +63,7 @@ export default class Input extends Component {
 
   render() {
     const {
-      className, inputContainerClassName, inputClassName, errorStyle, openScan, qr,
+      className, inputContainerClassName, inputClassName, errorStyle,
       valueLink: { error }, valueLink, dontDisplayError, inputCustomStyle,
       multiline, focusOnInit, disabled, readOnly, type, usd, ...rest
     } = this.props
@@ -96,12 +96,7 @@ export default class Input extends Component {
             })
           }
           { usd > 0 &&
-            <p styleName="rightEl" >{`~${usd}`}$</p>
-          }
-          { qr &&
-            <p styleName="rightEl qr" >
-              <i className="fas fa-qrcode" onClick={openScan} />
-            </p>
+            <p styleName="textUsd" >{`~${usd}`}$</p>
           }
         </div>
         {
