+++ shared/components/modals/WithdrawModal/WithdrawModal.js
@@ -18,7 +18,6 @@ import Tooltip from 'components/ui/Tooltip/Tooltip'
 import { FormattedMessage, injectIntl, defineMessages } from 'react-intl'
 import ReactTooltip from 'react-tooltip'
 import { isMobile } from 'react-device-detect'
-import QrReader from 'components/QrReader'
 
 // import isCoinAddress from 'swap.app/util/typeforce'
 import typeforce from 'swap.app/util/typeforce'
@@ -56,7 +55,6 @@ export default class WithdrawModal extends React.Component {
 
     this.state = {
       isShipped: false,
-      openScanCam: '',
       address: '',
       amount: '',
       minus: '',
@@ -269,29 +267,8 @@ export default class WithdrawModal extends React.Component {
       return typeforce.isCoinAddress[currency.toUpperCase()](address)
     }
 
-    openScan = () => {
-      const { openScanCam } = this.state
-
-      this.setState(() => ({
-        openScanCam: !openScanCam,
-      }))
-    }
-
-    handleError = err => {
-      console.error(err)
-    }
-
-    handleScan = data => {
-      if (data) {
-        this.setState(() => ({
-          address: data.includes(':') ? data.split(':')[1] : data,
-        }))
-        this.openScan()
-      }
-    }
-
     render() {
-      const { address, amount, balance, isShipped, minus, ethBalance, openScanCam,
+      const { address, amount, balance, isShipped, minus, ethBalance,
         isEthToken, exCurrencyRate, currentDecimals, error } = this.state
       const { name, data: { currency }, tokenItems, items, intl } = this.props
 
@@ -340,13 +317,6 @@ export default class WithdrawModal extends React.Component {
 
       return (
         <Modal name={name} title={`${intl.formatMessage(title.withdrowModal)}${' '}${currency.toUpperCase()}`}>
-          {openScanCam &&
-            <QrReader
-              openScan={this.openScan}
-              handleError={this.handleError}
-              handleScan={this.handleScan}
-            />
-          }
           <p styleName={isEthToken ? 'rednotes' : 'notice'}>
             <FormattedMessage
               id="Withdrow213"
@@ -369,13 +339,7 @@ export default class WithdrawModal extends React.Component {
                 </div>
               </Tooltip>
             </FieldLabel>
-            <Input
-              valueLink={linked.address}
-              focusOnInit pattern="0-9a-zA-Z:"
-              placeholder={`Enter ${currency.toUpperCase()} address to transfer`}
-              qr
-              openScan={this.openScan}
-            />
+            <Input valueLink={linked.address} focusOnInit pattern="0-9a-zA-Z:" placeholder={`Enter ${currency.toUpperCase()} address to transfer`} />
             {address && !this.addressIsCorrect() && (
               <div styleName="rednote">
                 <FormattedMessage
