+++ shared/components/modals/WithdrawModal/WithdrawModal.scss
@@ -145,9 +145,3 @@
     text-align: center;
   }
 }
-
-.qrCode {
-  position: absolute;
-  left: 97%;
-  top: 9px;
-}
