+++ shared/components/ui/DropDown/DropDown.js
@@ -5,13 +5,12 @@ import ClickOutside from 'react-click-outside'
 import Link from 'sw-valuelink'
 
 import cssModules from 'react-css-modules'
+import styles from './DropDown.scss'
 
 import toggle from 'decorators/toggle'
 import FieldLabel from 'components/forms/FieldLabel/FieldLabel'
 import Tooltip from 'components/ui/Tooltip/Tooltip'
 import Input from 'components/forms/Input/Input'
-
-import styles from './DropDown.scss'
 
 import closeBtn from './images/close.svg'
