--- shared/components/ui/DropdownMenu/DropdownMenu.scss
@@ -1,65 +1,0 @@
-.dropdownMenu {
-    position: static;
-    > button {
-        position: absolute;
-        right: 5px;
-        top: 50%;
-        transform: translateY(-50%);
-        cursor: pointer;
-        z-index: 111;
-        width: 20px;
-    }
-}
-
-.dropdownMenuItem {
-    border-bottom: 1px solid #f5f5f5;
-    &:last-child {
-        border: none;
-    }
-}
-
-.menu {
-    background: #FFFFFF;
-    box-shadow: 0px 4px 16px rgba(0,0,0,0.08), 0px 1px 4px rgba(0,0,0,0.08);
-    border-radius: 4px;
-    padding: 5px 10px;
-    width: 100px;
-    display: none;
-    button {
-        font-size: 10px;
-        margin: 0;
-        margin-bottom: 5px;
-    }
-    button[disabled] {
-        color: #8e9aa2;
-        background: #fff;
-        &:hover {
-            color: #8e9aa2;
-            background: #fff;
-        }
-    }
-}
-
-.menu.open {
-    display: block;
-}
-
-.walletControls {
-    position: absolute;
-    right: 12px;
-    top: -30px;
-    z-index: 111111;
-    padding: 0 5px;
-}
-
-.regular {
-    color: #000;
-}
-
-@media all and (min-width: 320px) and (max-width: 767px) {
-    .walletControls {
-        top: -30px;
-        padding: 0 5px;
-    }
-}
-
