+++ shared/pages/CreateWallet/CreateWallet.js
@@ -4,7 +4,6 @@ import CSSModules from 'react-css-modules'
 import styles from './CreateWallet.scss'
 
 import { connect } from 'redaction'
-import actions from 'redux/actions'
 
 import { FormattedMessage, injectIntl } from 'react-intl'
 import { withRouter } from 'react-router-dom'
@@ -17,6 +16,7 @@ import { localisedUrl } from 'helpers/locale'
 import check from './images/check'
 import FirstStep from './Steps/FirstStep'
 import SecondStep from './Steps/SecondStep'
+import ThirdStep from './Steps/ThirdStep'
 
 import { color } from './chooseColor'
 
@@ -30,13 +30,17 @@ const CreateWallet = (props) => {
   const [step, setStep] = useState(1)
   const [error, setError] = useState('Choose something')
 
-  const steps = [1, 2]
+  const steps = [1, 2, 3]
 
 
   const handleClick = () => {
     setError(null)
 
-    if (step !== 2) {
+    if (step === 2) {
+      setError('Choose something')
+    }
+
+    if (step !== 3) {
       reducers.createWallet.newWalletData({ type: 'step', data: step + 1 })
       return setStep(step + 1)
     }
@@ -47,20 +51,19 @@ const CreateWallet = (props) => {
     setError(null)
     if (!Object.values(currencies).includes(true) && step === 1) {
       setError('Choose something')
-
       return
     }
 
-    if (!secure.length && step === 2) {
+    if (step === 2 && eMail.length) {
+      if (!/.+@.+\.[A-Za-z]+$/.test(eMail)) {
+        setError('Invalid e-mail')
+        return
+      }
+    }
+
+    if (!secure.length && step === 3) {
       setError('Choose something')
       return
-    }
-    if (step === 1) {
-      Object.keys(currencies).forEach(el => {
-        if (currencies[el]) {
-          actions.core.markCoinAsVisible(el.toUpperCase())
-        }
-      })
     }
     handleClick()
   }
@@ -71,7 +74,7 @@ const CreateWallet = (props) => {
         <h2>
           <FormattedMessage
             id="createWalletHeader1"
-            defaultMessage="Создайте кошелек  в три простых шага?"
+            defaultMessage="Создайте кошелек  в три простых шага?" 
           />
         </h2>
         {isMobile &&
@@ -86,6 +89,7 @@ const CreateWallet = (props) => {
         <div>
           {step === 1 && <FirstStep error={error} onClick={validate} setError={setError} />}
           {step === 2 && <SecondStep error={error} onClick={validate} setError={setError} />}
+          {step === 3 && <ThirdStep error={error} onClick={validate} setError={setError} />}
         </div>
       </div>
     </div>
