+++ shared/pages/CreateWallet/CreateWallet.scss
@@ -4,6 +4,12 @@
   margin: 0 auto;
 }
 
+.notYet {
+  text-align: center;
+  padding-top: 5px;
+  color: #6144e5;
+}
+
 .wrapper h2 {
   width: 400px;
   margin-bottom: 60px;
@@ -13,9 +19,16 @@
 .formBody {
   margin: 0 16px;
 }
+
+.inputWrapper,
 .currencyChooserWrapper,
 .inLine {
   display: flex;
+}
+
+.inputWrapper {
+  justify-content: center;
+  margin: 50px 0;
 }
 
 .notMain {
@@ -119,8 +132,10 @@
 
 .currencyChooserWrapper {
   justify-content: space-between;
-  padding: 30px;
+}
 
+.currencyChooserWrapper {
+  padding: 30px;
 }
 
 .card {
@@ -136,15 +151,6 @@
   display: grid;
   height: 140px;
   width: 110px;
-
-  em {
-    position: absolute;
-    text-align: right;
-    padding-left: auto;
-    width: 120px;
-    font-size: 10px;
-    color: #9b58b6;
-  }
 
   b {
     font-weight: bolder;
@@ -162,10 +168,6 @@
   }
 }
 
-.cardDisabled {
-  opacity: 0.7 !important;
-}
-
 .secondStepInput {
   height: 50px;
   border: 1px solid rgba(0, 0, 0, 0.1);
@@ -179,35 +181,9 @@
   width: 140px;
 }
 
-@media screen and (max-width: 750px){
-  .secureSize {
-    width: 100% !important;
-    height: 95px !important;
-
-  }
-  .currencyChooserWrapperSecond{
-    overflow-y: auto;
-    display: block;
-    padding: 0;
-  }
-
-  .currencyChooserWrapperSecond{
-    height: 225px;
-    margin: 10px 0;
-  }
-
-  .card {
-    em {
-      position: relative;
-      width: 100%;
-      height: 0;
-    }
-  }
-}
-
 @media screen and (max-width: 700px){
   .card {
-    width: 85px;
+    width: 85px
   }
 }
 
@@ -220,12 +196,10 @@
 
 @media screen and (max-width: 480px) {
   .continue {
-    position: fixed;
-    z-index: 999999;
+    position: absolute;
     right: 21px;
-    width: 100%;
-    top: 95vh;
-    left: 0;
+    width: 89%;
+    top: 92vh;
   }
 
   .secondStepInput {
@@ -264,17 +238,7 @@
     margin: 0;
   }
 
-  .currencyChooserWrapper {
-    overflow-y: auto;
-    display: block;
-    padding: 0;
-  }
-
-  .currencyChooserWrapper {
-    height: 50%;
-  }
-
-  .second {
+  .third {
     height: 59px;
     margin-bottom: 0px;
   }
@@ -302,7 +266,7 @@
   }
 
   .steps {
-    width: 100px
+    width: 150px
   }
 
   .logo {
@@ -310,18 +274,13 @@
   }
 
   .listGroup {
-    margin: auto;
+    margin: auto 10px;
   }
 
   .secureSize,
   .card {
     height: 53px;
     width: 100%;
-
-    em {
-      position: absolute;
-      width: 70%;
-    }
   }
 
   .card {
@@ -349,6 +308,22 @@
     height: 65px;
   }
 
+  .inputWrapper {
+    display: block;
+    margin: 0;
+  }
+
+  .currencyChooserWrapper{
+    overflow-y: auto;
+    display: block;
+    padding: 0;
+  }
+
+  .inputWrapper,
+  .currencyChooserWrapper{
+    height: 265px;
+  }
+
   .thirdPageIcons {
     background: #f0eefd;
   }
@@ -370,7 +345,9 @@
   }
 
   .secureChooserWrapper,
+  .inputWrapper,
   .currencyChooserWrapper {
+    height: 50%;
     overflow-y: auto;
   }
 
@@ -387,8 +364,7 @@
 
 @media screen and (max-width: 320px){
   .continue {
-    top: 92vh;
-    width: 100%;
-    left: 0;
+    top: 91vh;
+    width: 87%;
   }
 }
