+++ shared/pages/CurrencyWallet/CurrencyWallet.scss
@@ -12,10 +12,6 @@
   justify-content: center;
 }
 
-.history {
-  display: table;
-}
-
 .subtitle {
   margin: 0;
   font-size: 18px;
