+++ shared/pages/History/Row/Row.scss
@@ -54,22 +54,20 @@
 }
 
 .confirm {
+  color: #8E9AA3;
   text-align: center;
   height: auto;
-  background: rgba(97, 68, 229, 0.1);
+  background: rgba(142, 154, 163, 0.1);
   font-weight: 800;
   font-size: 8px;
   line-height: 10px;
-  color: #6144E5;
   width: auto;
   text-transform: uppercase;
   padding: 3px 10px;
   border-radius: 2px;
   margin-left: 10px;
   position: relative;
-  top: -2px; a {
-    font-size: inherit;
-  }
+  top: -2px;
 }
 
 .directionHeading {
@@ -77,8 +75,8 @@
 }
 
 .unconfirmed {
-  color: #8E9AA3;
-  background: rgba(142, 154, 163, 0.1);
+  color: #ff354b;
+  background: rgba(255, 53, 75, 0.4);
   text-align: center;
   width: 200px;
 }
@@ -127,21 +125,4 @@ tr.historyRow {
   display: flex;
   align-items: center;
   padding: 15px;
-}
-
-
-@media all and (min-width: 320px) and (max-width: 570px) {
-  tr.historyRow {
-    height: 120px;
-    align-items: flex-start;
-  }
-  .amountUsd {
-    right: 15px;
-    bottom: 38px;
-  }
-  .amount {
-    right: 15px;
-    bottom: 15px;
-    top: auto;
-  }
 }
