+++ shared/pages/NewWallet/NewWallet.scss
@@ -1,0 +1,174 @@
+.newWallet {
+    margin: 30px 0;
+    font-family: 'Manrope-Bold',sans-serif;
+}
+
+.newWalletHeading {
+    font-weight: 800;
+    font-size: 40px;
+    line-height: 120%;
+    color: #1F2D48;
+}
+
+.newWalletTable {
+    width: 100%;
+}
+  
+
+.newWalletNav {
+    padding: 0;
+    margin-top: 30px;
+    margin-bottom: 70px;
+}
+
+.newWalletNavItem {
+    display: inline-block;
+    margin-right: 40px;
+}
+
+.newWalletNavItem.active {
+    a {
+        border-bottom: 3px solid #6144E5;
+        color: #1F2D48!important;
+    }
+}
+
+.newWalletNavItemLink {
+    cursor: pointer;
+    font-size: 15px;
+    line-height: 140%;
+    padding-bottom: 25px;
+    color: #8E9AA3!important;
+}
+
+.newWalletBalance {
+    margin-top: 40px;
+    width: 31%;
+}
+
+.yourAssets,
+.activity {
+   display: none;
+}
+
+.yourAssets.active,
+.activity.active {
+    display: block;
+}
+
+.newWalletContent {
+    display: flex;
+    justify-content: space-between;
+}
+
+.yourAssets,
+.activity {
+    background: #FFFFFF;
+    box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.08), 0px 1px 4px rgba(0, 0, 0, 0.08);
+    border-radius: 4px;
+    padding: 25px;
+    margin-top: 40px;
+    width: 66%;
+    button {
+        margin-top: 20px;
+    }
+}
+
+.yourAssetsHeading,
+.activityHeading{
+    color: #1F2D48;
+    font-size: 24px;
+    line-height: 33px;
+    font-weight: bold;
+    margin-bottom: 10px;
+}
+
+.yourAssetsDescr {
+    color: #8E9AA3;
+    font-size: 13px;
+    font-family: 'Manrope-Medium',sans-serif;
+    line-height: 150%;
+}
+
+.yourAssetsTable {
+    margin-top: 27px;
+}
+
+.yourBalance {
+    background: #FFFFFF;
+    box-shadow: 0px 4px 16px rgba(0, 0, 0, 0.08), 0px 1px 4px rgba(0, 0, 0, 0.08);
+    border-radius: 4px;
+    height: fit-content;
+}
+
+.yourBalanceTop {
+    padding: 25px;
+    position: relative;
+}
+.yourBalanceDescr {
+    color: #8E9AA3;
+    font-size: 13px;
+    font-weight: bold;
+}
+.yourBalanceValue {
+    display: flex;
+    align-items: end;
+    width: 145px;
+    justify-content: space-between;   
+    svg {
+        margin-top: 10px;
+    }
+    p {
+        font-size: 32px;
+        line-height: 44px;
+        color: #1F2D48;
+        font-weight: bold;
+    }
+    span {
+        font-size: 18px;
+        line-height: 25px;
+        color: #6144E5;
+        font-weight: bold;
+        margin-top: 5px;
+    }
+}
+
+.yourBalanceBottom {
+    position: relative;
+    padding: 25px;
+    display: flex;
+    justify-content: space-between;
+    button {
+        width: 48%;
+        cursor: pointer;
+        z-index: 1;
+    }
+    &:before {
+        position: absolute;
+        width: 100%;
+        height: 100px;
+        left: 0;
+        bottom: 0;
+        opacity: 0.1;
+        content: '';
+        background: linear-gradient(180deg, #6144E5 0%, rgba(97, 68, 229, 0) 100%);
+    }
+    height: 95px;
+}
+
+.yourBalanceCurrencies {
+    color: rgba(0, 0, 0, 0.1);
+    position: absolute;
+    background-color: white;
+    right: 25px;
+    top: 25px;
+    width: 45px;
+    justify-content: space-between;
+    display: flex;
+    span {
+        margin-top: 2px;
+    }
+    a {
+        cursor: pointer;
+    }
+}
