+++ shared/pages/NewWallet/components/NotityBlock/NotifyBock.js
