+++ shared/pages/Wallet/Row/Row.scss
