--- shared/pages/Wallet/components/NotityBlock/NotifyBlock.scss
@@ -1,147 +1,0 @@
-.notifyBlock {
-    height: 88px;
-    border-radius: 4px;
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding: 24px;
-    > div {
-        display: flex;
-        align-items: center;
-        button {
-            margin-left: 10px;
-            font-weight: bold;
-        }
-    }
-
-}
-
-.notifyBlockSaveKeys {
-    background: linear-gradient(265.47deg, #6144E5 0%, #7359EA 100%);
-    > div {
-        color: #6144E5;
-    }
-}
-
-.notifyBlockSignUp {
-    background: linear-gradient(265.47deg, #FF6B00 0%, #F88736 100%);
-    > div {
-        color: #F1771F;
-    }
-}   
-
-.notifyBlockIcon {
-    background: rgba(255, 255, 255, 0.1);
-    width: 40px;
-    height: 40px;
-    border-radius: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-right: 7px;
-}
-
-.notifyBlockDescr {
-    margin-left: 10px;
-    color: #fff;
-    font-family: 'Manrope-Bold',sans-serif;
-    p {
-        margin-bottom: 0;
-    }
-    p:last-child {
-        opacity: 0.8;
-        font-size: 13px;
-        font-family: 'Manrope-Medium',sans-serif;
-    }
-}
-
-@media all and (min-width: 768px) and (max-width: 992px) {
-    .notifyBlock {
-        height: max-content;
-        > div:first-child{
-            width: 60%;
-        }
-        button {
-            font-size: 11px;
-        }
-    }
-    .notifyBlockDescr {
-        p {
-            font-size: 12px;
-            margin-bottom: 0px;
-        }
-        p:last-child {
-            font-size: 12px;
-        }
-    }
-    .notifyBlockIcon {
-        align-self: flex-start;
-        min-width: 40px;
-    }
-}
-@media all and (min-width: 320px) and (max-width: 767px) {
-    .notifyBlock {
-        flex-direction: column;
-        height: max-content;
-        align-items: flex-start;
-        > div {
-            align-items: flex-start;
-            button {
-                font-size: 11px;
-                margin: 0;
-            }
-            button:last-child {
-                margin-left: 10px;
-            }
-        }
-        > div:last-child {
-            width: 100%;
-            margin-top: 20px;
-            display: flex;
-            justify-content: center;
-        }
-    }
-    .notifyBlockIcon {
-        min-width: 40px;
-    }
-}
-
-@media all and (min-width: 571px) and (max-width: 767px) {
-    .notifyBlockDescr {
-        p {
-            font-size: 13px;
-            margin-bottom: 0px;
-        }
-        p:last-child {
-            font-size: 12px;
-        }
-    }
-}
-@media all and (min-width: 320px) and (max-width: 570px) {
-    .notifyBlockDescr {
-        p {
-            font-size: 13px;
-            margin-bottom: 4px;
-        }
-        p:last-child {
-            font-size: 12px;
-        }
-    }
-}
-
-@media all and (min-width: 320px) and (max-width: 340px) {
-    .notifyBlock {
-        > div {
-            align-items: flex-start;
-            button {
-                font-size: 10px;
-                margin: 0;
-            }
-            button:last-child {
-                margin-left: 10px;
-            }
-        }
-    }
-}
-
-
