+++ shared/routes/index.js
@@ -6,12 +6,12 @@ import { localisePrefix } from 'helpers/locale'
 
 import SwapComponent    from 'pages/Swap/Swap'
 import Home             from 'pages/Home/Home'
-import OldWallet        from 'pages/OldWallet/Wallet'
+import Wallet           from 'pages/Wallet/Wallet'
 import History          from 'pages/History/History'
 import CreateWallet     from 'pages/CreateWallet/CreateWallet'
 import NotFound         from 'pages/NotFound/NotFound'
 import About            from 'pages/About/About'
-import Wallet           from 'pages/Wallet/Wallet'
+import NewWallet        from 'pages/NewWallet/NewWallet'
 import Currency         from 'pages/Currency/Currency'
 import PartialClosure   from 'pages/PartialClosure/PartialClosure'
 import CurrencyWallet   from 'pages/CurrencyWallet/CurrencyWallet'
@@ -35,7 +35,7 @@ const routes = (
       <Route path={`${localisePrefix}${links.exchange}`} component={PartialClosure} />
 
       <Route path={`${localisePrefix}${links.aboutUs}`} component={About} />
-      <Route path={`${localisePrefix}${links.wallet}`} component={Wallet} />
+      <Route path={`${localisePrefix}${links.newWallet}`} component={NewWallet} />
       <Route path={`${localisePrefix}${links.history}`} component={History} />
       <Route path={`${localisePrefix}${links.createWallet}`} component={CreateWallet} />
 
@@ -45,7 +45,6 @@ const routes = (
       <Route exact path={`${localisePrefix}${links.home}`} component={PartialClosure} />
 
       <Route path={`${localisePrefix}${links.currencyWallet}`} component={Wallet} />
-      <Route path={`${localisePrefix}${links.oldWallet}`} component={OldWallet} />
       <Route path={`${localisePrefix}${links.home}:currency`} component={Currency} />
